add parameter of image and enable the context to refresh it's token.

+++ cloud/src/test/java/com/datayes/cloud/openstack/TenantManagerTest.java
@@ -18,12 +18,12 @@ import static org.junit.Assert.assertEquals;
 public class TenantManagerTest {
 
     private static final Logger log = LoggerFactory.getLogger(TenantManagerTest.class);
-    private String identityServiceUrl = "http://10.20.112.226:5000/v2.0/tokens";
+    private String identityServiceUrl = "http://10.20.112.65:5000/v2.0/tokens";
     private OpenstackContext openstackContext;
 
     @Before
     public void setUp() throws Exception {
-        openstackContext = new OpenstackContext(identityServiceUrl, "admin", "aaaaaa", "admin");
+        openstackContext = new OpenstackContext(identityServiceUrl, "admin", "admin", "admin");
 
 
     }
@@ -40,11 +40,20 @@ public class TenantManagerTest {
     @Test
     public void testCreateTenant() throws Exception {
         TenantManager tenantManager = new TenantManager(openstackContext);
-        tenantManager.deleteTenant("tenant1");
+
         List<Tenant> tenants = tenantManager.listTenants();
-        int oldSize = tenants.size();
+        boolean exists = false;
+        for (Tenant t : tenants){
+            if (t.getName().equals("tenant1")){
+                tenantManager.deleteTenant("tenant1");
+                exists = true;
+                break;
+            }
+        }
+        int oldSize = tenants.size()-(exists?1:0);
         Tenant tenant = tenantManager.createTenant("tenant1", "tenant1 desc", true);
         assertEquals("tenant1", tenant.getName());
+
         tenants = tenantManager.listTenants();
         assertEquals(oldSize + 1, tenants.size());
     }
@@ -55,18 +64,14 @@ public class TenantManagerTest {
         tenantManager.deleteTenant("tenant1");
         Tenant tenant = tenantManager.createTenant("tenant1", "tenant1 desc", true);
         System.out.println(tenant.getId());
-        OpenstackContext tenant1Context = new OpenstackContext(identityServiceUrl, "admin", "aaaaaa", "tenant1");
-        NetworkManager networkManager = new NetworkManager(tenant1Context);
-        Network network = networkManager.createNetwork("network1", false);
-        System.out.println(network);
-        Subnet subnet = networkManager.createSubnet(network);
-        System.out.println(subnet);
+        OpenstackContext tenant1Context = new OpenstackContext(identityServiceUrl, "admin", "admin", "tenant1");
+
 //        networkManager.createNetwork("network2", false);
         ComputeManager computeManager = new ComputeManager(tenant1Context);
         StorageManager storageManager = new StorageManager(tenant1Context);
         Volume volume = storageManager.createVolume(new Volume("v1", 1));
         Server server = new Server("server1");
-        Server result = computeManager.createServer(server, volume.getId());
+        Server result = computeManager.createServer(server, volume.getId(),"TestVM");
         System.out.println(volume);
         System.out.println(result);
     }
@@ -79,7 +84,7 @@ public class TenantManagerTest {
 
     @Test
     public void testListServers() throws Exception {
-        OpenstackContext tenant1Context = new OpenstackContext(identityServiceUrl, "admin", "aaaaaa", "tenant1");
+        OpenstackContext tenant1Context = new OpenstackContext(identityServiceUrl, "admin", "admin", "tenant1");
         ComputeManager computeManager = new ComputeManager(tenant1Context);
         List<Server> servers = computeManager.listServers();
         for (Server server : servers) {
@@ -89,12 +94,12 @@ public class TenantManagerTest {
 
     @Test
     public void testListVolumes() throws Exception {
-        OpenstackContext tenant1Context = new OpenstackContext(identityServiceUrl, "admin", "aaaaaa", "tenant1");
+        OpenstackContext tenant1Context = new OpenstackContext(identityServiceUrl, "admin", "admin", "tenant1");
         StorageManager storageManager = new StorageManager(tenant1Context);
         List<Volume> volumes = storageManager.listVolumes();
         for (Volume volume : volumes) {
             System.out.println(volume);
-            storageManager.deleteVolume(volume.getId());
+            //storageManager.deleteVolume(volume.getId());
         }
     }
